Make failing test pass :)
Actually, this commit here is a rewrite of Array.each (and family) so that we don't have to worry about saving and restoring the selections when looping (and transforming) a collection of elements.

Some commands might be broken as a result of this change (those that are not tested!), so some tests are in order before release.

Also, benchmarking shows some slowdown in Fireworks CS4 and CS5 that I need to investigate:

Old tests and code
CS3: 759 ms
CS4: 1359 ms
CS5: 1040 ms

New tests and code
CS3: 730 ms
CS4: 1690 ms
CS5: 1340 ms